+++ src/Language/FCL/Orphans.hs
@@ -1,0 +1,19 @@
+{-# options_ghc -fno-warn-orphans #-}
+
+module Language.FCL.Orphans () where
+
+import Protolude
+
+import Data.List.NonEmpty (NonEmpty)
+import qualified Data.List.NonEmpty as NonEmpty
+import Data.Serialize (Serialize(..))
+import qualified Data.Serialize as Serialize
+import Test.QuickCheck (Arbitrary(..))
+
+
+instance Serialize a => Serialize (NonEmpty a) where
+  get = NonEmpty.fromList <$> Serialize.get
+  put = Serialize.put . NonEmpty.toList
+
+instance Arbitrary a => Arbitrary (NonEmpty a) where
+  arbitrary = (:|) <$> arbitrary <*> arbitrary