+++ tests/Main.hs
@@ -65,6 +65,7 @@ main :: IO ()
 main = do
   swaggerTests <- testSpec "Swagger test" $ TestSwagger.swaggerTest
   fclTests <- suite
-  defaultMain $ testGroup "All tests" [
-    swaggerTests,
-    fclTests]
+  defaultMain $ testGroup "All tests"
+    [ fclTests
+    , swaggerTests
+    ]